Fixed problem with - adding, edditing, deleaning books - small fixes related to the views - added validations to Author and to Book classes

+++ Controllers/AuthorsController.cs
@@ -107,8 +107,11 @@ namespace LibraryManagerApp.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var author = await _context.Authors.FindAsync(id);
-            _context.Authors.Remove(author);
-            await _context.SaveChangesAsync();
+            if (author != null)
+            {
+                _context.Authors.Remove(author);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
