+++ Controllers/BooksController.cs
@@ -111,8 +111,11 @@ public class BooksController : Controller
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
         var book = await _context.Books.FindAsync(id);
-        _context.Books.Remove(book);
-        await _context.SaveChangesAsync();
+        if (book != null)
+        {
+            _context.Books.Remove(book);
+            await _context.SaveChangesAsync();
+        }
         return RedirectToAction(nameof(Index));
     }
